requirements.txt
flask==3.0.0
flask_sqlalchemy==3.1.1
gunicorn==20.1.0
requests==2.31.0
docker==6.1.3
python-dotenv==1.0.0
pyjwt==2.8.0
bcrypt==4.1.2
pyyaml==6.0.1